-Added Real World Scale option. Using this, the models will export into a real-world scale! (Currently only works on Lightwave and OBJ exporters.) -Lightwave Exporter:  -Will now ask if you want to overwrite model files.  -Scene files will no longer be generated if the object doesn't export.  -Added Always Export Scene option. When enabled, this will export the scene file, even if the object doesn't export for some reason.  -All exported items obey the Real World Scale option. -OBJ Exporter:  -Now exports attached models! (Helms, Weapons and Shoulders)  -Now obeys the Preserve Directory option!  -Image filenames now work just like the Lightwave exporter's image filenames.  -All exported items obey the Real World Scale option. -DEV WORK:  -Massive, and I do mean MASSIVE 64-bit updates.  -Moved Mesh and Slot names to the modelexport.h  -Removed fn from GetM2TextureName, as it wasn't needed.  -Fixed MakeModelFaceForwards, and changed the 2nd option to flip the Z axis, instead of the X axis. If your model comes out backwards when using this, set flipZ to true.

+++ trunk/src/database.h
@@ -217,9 +217,9 @@ public:
 	static const size_t HairType = 3;
 	static const size_t UnderwearType = 4;
 
-	Record getByParams(unsigned int race, unsigned int gender, unsigned int type, unsigned int section, unsigned int color, unsigned int npc);
-	int getColorsFor(unsigned int race, unsigned int gender, unsigned int type, unsigned int section, unsigned int npc);
-	int getSectionsFor(unsigned int race, unsigned int gender, unsigned int type, unsigned int color, unsigned int npc);
+	Record getByParams(size_t race, size_t gender, size_t type, size_t section, size_t color, size_t npc);
+	int getColorsFor(size_t race, size_t gender, size_t type, size_t section, size_t npc);
+	int getSectionsFor(size_t race, size_t gender, size_t type, size_t color, size_t npc);
 };
 
 
@@ -690,7 +690,7 @@ public:
 	static const size_t Filename = 20;		// string. an index offset to the filename.
 
 	Record getByFilename(wxString fn);
-	Record getByNPCID(unsigned int id);
+	Record getByNPCID(size_t id);
 
 };
